ci-platform-jjb: try LAVA job in multiple labs
Currently, LAVA jobs are only submitted to a single lab.  This patch
adds the ability look for a matching device-type in a list of
available labs, and submit the job to the first lab with an available
device type.

To accomplish this, the job submission part is removed from
run-test-short.sh, which now only generates the LAVA YAML job.  A new
lava-jobs-submit.sh script is added which looks for an available
device-type in a list of LAVA labs, and submits the job there.

NOTES:
- run-test-short.sh should probably be renamed, but was left
  for now to limit the diff

Change-Id: I95086d89a95f39e2b59fec8a8afada70806c52e5
Signed-off-by: Kevin Hilman <khilman@baylibre.com>