run-test-prepare: fix device environment overrides
Machines can override the default environment settings in their
hwtest.short.environment files.  However, the grep/sed procesing done
does not ignore lines that are commented out, causing an extra '#' to
be part of the variable name.

To fix, ensure the grep for the variable name matches the beginning of
the line, so commented lines will be ignored.

Change-Id: Ibf4281c3126c95670346bee74043d670e6d58654
Signed-off-by: Kevin Hilman <khilman@baylibre.com>